# docs/api_swagger.yaml
basePath: /api
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  schema.Account:
    properties:
      application:
        $ref: '#/definitions/schema.Application'
      application_id:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      roleApplications:
        items:
          $ref: '#/definitions/schema.RoleApplication'
        type: array
      status:
        type: boolean
      updatedAt:
        type: string
      user_code:
        type: string
    type: object
  schema.ActivityLog:
    properties:
      application:
        type: string
      body:
        type: string
      client:
        type: string
      createdAt:
        type: string
      duration:
        type: string
      id:
        type: integer
      method:
        type: string
      path:
        type: string
      path_op:
        type: string
      referrer:
        type: string
      request_id:
        type: string
      source:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/schema.User'
      user_id:
        type: string
    type: object
  schema.Application:
    properties:
      accounts:
        items:
          $ref: '#/definitions/schema.Account'
        type: array
      app_package:
        type: string
      app_package_class:
        type: string
      asset_apk:
        type: string
      asset_icon:
        type: string
      code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      updated_note:
        type: string
      updatedAt:
        type: string
      version:
        type: string
    type: object
...
  schema.Article:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      status:
        type: boolean
      title:
        type: string
      updatedAt:
        type: string
      user_id:
        type: string
    type: object
  schema.Company:
    properties:
      address:
        type: string
      code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      domain:
        type: string
      fax:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      sector:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  schema.Division:
    properties:
      code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      estate:
        $ref: '#/definitions/schema.Estate'
      estate_code:
        type: string
      gangs:
        items:
          $ref: '#/definitions/schema.Gang'
        type: array
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  schema.Employee:
    properties:
      address:
        type: string
      code:
        type: string
      company:
        $ref: '#/definitions/schema.Company'
      company_code:
        type: string
      cost_center:
        type: string
      created_by:
        type: string
      createdAt:
        type: string
      deleted:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      department:
        type: string
      email:
        type: string
      expired_time:
        type: integer
      expiredAt:
        type: string
      fiscal_period:
        type: string
      fiscal_year:
        type: string
      fullname:
        type: string
      grade_rate:
        type: string
      harvester_code:
        type: string
      id:
        type: string
      job_pos:
        type: string
      mandt:
        type: string
      nickname:
        type: string
      nik:
        type: string
      office_number:
        type: string
      phone_number:
        type: string
      picture:
        type: string
      status:
        type: boolean
      type:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/schema.User'
      user_code:
        type: string
      username:
        type: string
    type: object
  schema.Estate:
    properties:
      code:
        type: string
      company:
        $ref: '#/definitions/schema.Company'
      company_code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      divisions:
        items:
          $ref: '#/definitions/schema.Division'
        type: array
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  schema.Gang:
    properties:
      code:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      division:
        $ref: '#/definitions/schema.Division'
      division_code:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: boolean
    type: object
  schema.GangResponse:
    properties:
      code:
        type: string
      division_code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  schema.MetaAccount:
    properties:
      application:
        $ref: '#/definitions/schema.MetaApp'
      id:
        type: string
      roleApplication:
        items:
          $ref: '#/definitions/schema.MetaRoleApp'
        type: array
      userCode:
        type: string
    type: object
  schema.MetaActivities:
    properties:
      application:
        type: string
      client:
        type: string
      createdAt:
        type: string
      duration:
        type: string
      id:
        type: integer
      method:
        type: string
      path:
        type: string
      pathOp:
        type: string
      referrer:
        type: string
      reqBody:
        type: string
      requestId:
        type: string
      source:
        type: string
      status:
        type: string
      userId:
        type: string
    type: object
  schema.MetaApp:
    properties:
      appPackage:
        type: string
      appPackageClass:
        type: string
      assetApk:
        type: string
      assetIcon:
        type: string
      description:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      updatedNote:
        type: string
      version:
        type: string
    type: object
  schema.MetaCompany:
    properties:
      address:
        type: string
      code:
        type: string
      description:
        type: string
      domain:
        type: string
      fax:
        type: string
      id:
        type: string
      name:
        type: string
      phoneNumber:
        type: string
      sector:
        type: string
    type: object
  schema.MetaDivision:
    properties:
      code:
        type: string
      description:
        type: string
      gang:
        items:
          $ref: '#/definitions/schema.MetaGang'
        type: array
      id:
        type: string
      name:
        type: string
    type: object
  schema.MetaEmployee:
    properties:
      address:
        type: string
      code:
        type: string
      company:
        $ref: '#/definitions/schema.MetaCompany'
      companyCode:
        type: string
      costCenter:
        type: string
      department:
        type: string
      email:
        type: string
      employeeType:
        type: string
      expired:
        type: integer
      fiscalPeriod:
        type: string
      fiscalYear:
        type: string
      fullName:
        type: string
      gradeRate:
        type: string
      harvesterCode:
        type: string
      id:
        type: string
      jobPos:
        type: string
      mandt:
        type: string
      nickName:
        type: string
      nik:
        type: string
      officeNumber:
        type: string
      phoneNumber:
        type: string
      picture:
        type: string
      userCode:
        type: string
      username:
        type: string
    type: object
  schema.MetaEstate:
    properties:
      code:
        type: string
      company:
        $ref: '#/definitions/schema.MetaCompany'
      description:
        type: string
      division:
        items:
          $ref: '#/definitions/schema.MetaDivision'
        type: array
      id:
        type: string
      name:
        type: string
    type: object
  schema.MetaGang:
    properties:
      code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  schema.MetaRoleApp:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  schema.MetaRoles:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  schema.MetaUser:
    properties:
      accounts:
        items:
          $ref: '#/definitions/schema.MetaAccount'
        type: array
      activityLogs:
        items:
          $ref: '#/definitions/schema.MetaActivities'
        type: array
      code:
        type: string
      createdAt:
        type: string
      email:
        type: string
      employees:
        items:
          $ref: '#/definitions/schema.MetaEmployee'
        type: array
      estates:
        items:
          $ref: '#/definitions/schema.MetaEstate'
        type: array
      id:
        type: string
      isActive:
        type: boolean
      password:
        type: string
      roles:
        items:
          $ref: '#/definitions/schema.MetaRoles'
        type: array
      updatedAt:
        type: string
      username:
        type: string
    type: object
  schema.MsgResponse:
    properties:
      msg:
        type: string
    type: object
...
  schema.RequestGang:
    properties:
      code:
        type: string
      division_code:
        type: string
      name:
        type: string
    required:
    - code
    - division_code
    - name
    type: object
  schema.RequestGangUpdate:
    properties:
      code:
        type: string
      delete:
        type: boolean
      division_code:
        type: string
      name:
        type: string
      status:
        type: boolean
    required:
    - code
    - delete
    - division_code
    - name
    - status
    type: object
  schema.Role:
    properties:
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
  schema.RoleApplication:
    properties:
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
  schema.SignUpJsonSchema:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        items:
          type: string
        type: array
      username:
        type: string
    required:
    - email
    - password
    - role
    - username
    type: object
  schema.User:
    properties:
      accounts:
        items:
          $ref: '#/definitions/schema.Account'
        type: array
      activityLogs:
        items:
          $ref: '#/definitions/schema.ActivityLog'
        type: array
      articles:
        items:
          $ref: '#/definitions/schema.Article'
        type: array
      code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      divisions:
        items:
          $ref: '#/definitions/schema.Division'
        type: array
      email:
        type: string
      employees:
        items:
          $ref: '#/definitions/schema.Employee'
        type: array
      id:
        type: string
      roles:
        items:
          $ref: '#/definitions/schema.Role'
        type: array
      status:
        type: boolean
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: Fajar soegi
    url: http://www.swagger.io/support
  description: This is User Service Test.
  termsOfService: http://swagger.io/terms/
  title: Swagger User Service
  version: "1.0"
paths:
...
  /gang:
    get:
      consumes:
      - application/json
      description: Get Gang All
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schema.GangResponse'
            type: array
      summary: Get All Gang
      tags:
      - (D) Gang
    patch:
      consumes:
      - application/json
      description: Update Gang
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Company Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestGangUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Update Gang
      tags:
      - (D) Gang
    post:
      consumes:
      - application/json
      description: Create Gang
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Estate Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestGang'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Create Gang
      tags:
      - (D) Gang
  /gang/{id}:
    get:
      consumes:
      - application/json
      description: Get Gang ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Division ID OR Division Code
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.GangResponse'
      summary: Get Gang By ID
      tags:
      - (D) Gang
  /user_new:
    post:
      consumes:
      - application/json
      description: User New
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Login Schema '
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/schema.SignUpJsonSchema'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: User New
      tags:
      - User Credential
  /user_profile:
    get:
      consumes:
      - application/json
      description: User Profile All
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schema.User'
            type: array
      summary: User Profile All
      tags:
      - User Credential
  /user_profile/{id}:
    get:
      consumes:
      - application/json
      description: User Profile
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: USER ID OR USER Code
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MetaUser'
      summary: User Profile
      tags:
      - User Credential
swagger: "2.0"